added a lot of comments to explain Link and examples
